Fixing the Procfile for gunicorn

+++ src/chit_main_app/management/commands/fix_pc.py
@@ -36,4 +36,5 @@ class Command(BaseCommand):
 
                     for x in delete_list:
                         print "Deleting: ", x.id
-                        x.delete()
+                        if x.id:
+                            x.delete()
